[DBC] use floats for item enchantment data
Missed in 6c75c80f06 because of the header name. Reduces binary size by
~65KiB.

+++ engine/dbc/spell_item_enchantment.hpp
@@ -21,7 +21,7 @@ struct item_enchantment_data_t
   unsigned    ench_type[3];        // item_enchantment
   int         ench_amount[3];
   unsigned    ench_prop[3];        // item_mod_type
-  double      ench_coeff[3];       // item enchant scaling multiplier for data table
+  float       ench_coeff[3];       // item enchant scaling multiplier for data table
   unsigned    id_spell;            // reverse mapped spell id for this enchant
   const char* name;
